refactor: complete ADR-061 Phase 3's per-artifact pipeline
Phase 3's target layout is now real rather than a contracts-only stub, and the
one edge its own status note named as the blocker is closed.

**The blocker first.** `service_compare_pipeline` was the last service module
without a `workflows` owner, held there by `prepare_embedded_build_source` /
`attach_evidence_metrics` living in `cli_buildsource`. Those, and the
`diff_embedded_build_source` they drive, are now
`buildsource/evidence_report.py`. The module deliberately owns no output
stream: it renders the ADR-028 D7 coverage, by-side and capability report as
`list[str]` and hands the lines to an optional `on_output` sink. That replaces
a `quiet: bool` for the same reason the `embed_build_source` move did -- a
quiet flag is only meaningful to a caller that has a stream, and
`run_compare_request` never did; it had to pass `quiet=True` forever to
suppress writes to a stream it does not own. With a sink, the engine default
(no sink, no output) is the correct one, and the CLI adapter decides its sink
writes to stderr so `--format json` stdout stays parseable.

The error contract is preserved rather than tidied. `resolve_side_pack` raises
`SnapshotError` for an unloadable pack -- operational, not usage: the command
line was well-formed and the pack's bytes were not -- and only the CLI adapter
renders that as `ClickException` (exit 1). Collapsing it into a `UsageError`
would tell a CI consumer the invocation was wrong when the data was.
`tests/test_evidence_report_contract.py` was written before the move; five of
its six cases passed against the pre-move code, pinning both exit codes, the
stdout/stderr split and the typed API's silence.

`evidence_policy`'s only `click` use was one echo function, so it became
`evidence_metrics_lines`. Two `ENGINE_CLI_BOUNDARY_ALLOWLIST` entries are now
stale (10 -> 8); the gate's own staleness test proves both closures rather
than asserting them.

**Then the pipeline.** `service_input_resolution.py` split along the seam it
already had: `workflows/artifact/resolve.py` decides a plan without running
it, `workflows/artifact/execute.py` runs one and reports what it achieved.
Three of Phase 3's acceptance criteria are structural properties of that split
rather than claims about it -- extraction happens once per artifact because
`resolve_side_snapshot` is the single entry point every front end reaches;
resource lifetimes cover execution because the L2 seed's cleanups drain
between the header parse and the embed; achieved depth is a result fact on
`SideResolution` rather than a frontend guess. The old module remains a
delegating facade, and production callers were repointed at the owners.

Tests that patched the old locations were repointed to the implementation
owners, not left to bind a name nothing reads. The repointing was verified
non-vacuous by reverting one patch target and watching it fail.

`architecture/modules.yaml` gains owners for the modules the migrated package
now imports; `check_architecture.py` and the boundary gate are both clean, and
mypy has no non-stub errors.
